+++ templates/video.md.js
@@ -1,6 +1,6 @@
 md(`
-<iframe title="About / Volcanic International"
-src="https://player.vimeo.com/video/212780263?dnt=1&amp;app_id=122963"
-width="1165" height="655"
-allow="autoplay; fullscreen; picture-in-picture; clipboard-write"></iframe>
+<a href="https://vimeo.com/212780263" target="_blank" rel="noopener noreferrer">
+<img src="https://i.vimeocdn.com/video/637934146-1d2216c41af89481288201476b8cea1835c433a3f8f0b2ec67e97c59b0cf2f43-d_640"
+alt="About / Volcanic International" width="640" height="360">
+</a>
 `)